Roll back to working state. (#780)
* Revert "Merge pull request #776 from Sinestro38/master"

This reverts commit b4c7ea151275cb75bd75e5c2d35f36abf8679fd8, reversing
changes made to 140277d10ca63b541acb481ecaf899f91c846e5b.

* Revert "remove release notes md"

This reverts commit 25313f41e8350ef90153bef0f03c9d576d34b1dd.

* Revert "Merge remote-tracking branch 'forkB/fix_warnings'"

This reverts commit d78931498fd4ee6d06ffba9271f2d948827c58fb, reversing
changes made to fe3a96e24a3001b76547a36bbe3301f8e5a8a255.

* Revert "Merge pull request #772 from Sinestro38/fix_license"

This reverts commit 140277d10ca63b541acb481ecaf899f91c846e5b, reversing
changes made to 7131f9e4d2d289e51f9705161b29c45159da1921.

* Revert "fix ubuntu version to 20.04"

This reverts commit 803622aa2f59983b81086dc07d562aca0c955f2b.

* Revert "[debug] get ubuntu  version of runner"

This reverts commit 8002690bcba9c4bdf28c961f38d2a89cd38f14dc.

* Revert "update ubuntu github runner from 18.04 to latest"

This reverts commit 471573ae6883f767efe61d20f5fa7a88de60c012.

* Revert "update > to ~"

This reverts commit fe3a96e24a3001b76547a36bbe3301f8e5a8a255.

* Revert "update cirq dependency news"

This reverts commit a44ab2659fde41daf73f1abbcca56aa34d78661d.

* Revert "Update release.md"

This reverts commit a7a67c8afc5f977a8347be0a78001c4c9fbd772a.

* Revert "add notes"

This reverts commit c7f729a01527bb065bdc5a0e7b1ea43e1330f4ba.

* Revert "update version num"

This reverts commit f933b6925b7ef673c78dbbb750e0a75e3c70dcd3.

* Revert "Update release.md"

This reverts commit c2ab0d5e08af1eed85f97e8ab5794bbbc7c20833.

* Revert "Create release.md"

This reverts commit ac7722ad7da4e529da40e0c2baccdaadf9c2c056.

* Revert "Merge branch 'master' into master"

This reverts commit 7131f9e4d2d289e51f9705161b29c45159da1921, reversing
changes made to 2c1e7fa4778475040c149b23e2b2f661007a1f8e.

* Revert "Upgrade bazel version 5.3.0 and fix some typo in tf version (#755)"

This reverts commit 2c1e7fa4778475040c149b23e2b2f661007a1f8e.

* Revert "Merge pull request #3 from jaeyoo/pavan/remove_compute_small"

This reverts commit 79aae25dc5db7c7e1ae2a83c23a0fb88af2d19c4, reversing
changes made to 46aa1f0eb43473967f81e14b00a706adefc37482.

* Revert "Fix lint"

This reverts commit 20815bc5f5977f189f467d5726f101b82f1498f1.

* Revert "Fix lint"

This reverts commit 47e6aaeefffa7d8ce66ea632998dceb2a15be28c.

* Revert "Update release/BUILD with `if_cuda_is_configured`"

This reverts commit f2cd2552b001004bae0f3d843860e5485b7443a2.

* Revert "add :tfq_simulate_ops_cuquantum_py target to release BUILD"

This reverts commit 3e8a9a42d16e1b092d29db1af623c2ff8609baf8.

* Revert "remove computeSmall from cuda and cuquantum kernels"

This reverts commit 5cb7fa153f694e4ae1d1eb94c69b78e5b0d88e4b.

* Revert "Merge pull request #2 from jaeyoo/cuquantum_support"

This reverts commit 46aa1f0eb43473967f81e14b00a706adefc37482, reversing
changes made to 2a5fd073deecc9c9a07cf2d85e7befb3f3554fcb.

* Revert "Merge pull request #1 from jaeyoo/cuda_support"

This reverts commit 2a5fd073deecc9c9a07cf2d85e7befb3f3554fcb, reversing
changes made to b0dd7451567fa1a7ad6f0907320344617892c9ea.

* Revert "Merge branch 'master' into cuda_support"

This reverts commit f19298ccb5bdcbee27b931f43690cad228420f57, reversing
changes made to 4f6290acab99f65733c36aff4f628c1b11ae42a9.

* Revert "Add cuQuantum test (but failed at numeric diff test)"

This reverts commit f97c496bfd939b2ec5534718618be76274b76e20.

* Revert "Fix ComputeSmall and enable it."

This reverts commit 4f6290acab99f65733c36aff4f628c1b11ae42a9.

* Revert "Fix tfq_simulate_ops_cuda op test."

This reverts commit 66442482a21fc7139fb2cdc69a36fd2f5e9c207f.

* Revert "Temporarily turn on ComputeLarge always"

This reverts commit 23dd880224b209430dfda4c3a88a351c64dce28a.

* Revert "Upgrade TF version from 2.7 to 2.11 (#749)"

This reverts commit 20cfde5eb2f6188d35bab1375c98572ae4571b20.

* Revert "TFQ simulate expectation CUDA op"

This reverts commit f063417f4b08c6b430c8822bc9cb8ad8a7819ff9.

* Revert "Update ci_install bazel version, removed duplicated code lines."

This reverts commit 1317edf584cf978a106ecb657faf3d3fdef67402.

* Revert "Fix typo"

This reverts commit a765f7ff84e2b7edb19e324b8d6a33171a7d5984.

* Revert "Upgrade seaborn version to 0.12.0"

This reverts commit 4e7c22dedc2e231d7eab3dd4cd3f0235444e51e0.

* Revert "Fix tutorial errors"

This reverts commit c185891066ce453dcac56fc19d172bd71150a031.

* Revert "Fix tutorial error"

This reverts commit c5df061088fc7d69fde7f8f695e0a43fd2a32213.

* Revert "Upgrade the latest gym for notebook tutorial."

This reverts commit 9b94ebf2e4ff4cbabacf03bd52eebab4327409f0.

* Revert "Format optimizers"

This reverts commit 39a811d1a2ba8e854f800ba23157a7cc017a8cf4.

* Revert "Format"

This reverts commit a22b98fa4fef34a4df9174e33d2c455b675337c4.

* Revert "Upgrade to TF v2.11"

This reverts commit 26fa2dfda35ab64585d03e20def1dfd42031b123.

* Upgrade TF version from 2.7 to 2.11 (#749)

* Upgrade bazel version to 5.1.0 for building with TFv2.11.0

* Upgrade to TF v2.11

* Bump up the version of numpy to v1.24.2

* Format

* Fix tf.Variable inside tf.while_loop issues inside optimizers

* Format optimizers

* Upgrade the latest gym for notebook tutorial.

* Fix tutorial error

* Fix tutorial errors

* Upgrade seaborn version to 0.12.0

* Fix typo

* Remove debug print

* Update ci_install bazel version, removed duplicated code lines.

* Upgrade bazel version 5.3.0 and fix some typo in tf version (#755)

* fix compatability issues.

* re-trigger kokoro

* re-trigger kokoro

* re-trigger kokoro

* re-trigger kokoro

* re-trigger kokoro

* Lower tolerances.

* Bump CI python version.

---------

+++ tensorflow_quantum/core/ops/math_ops/tfq_inner_product_grad.cc
@@ -398,13 +398,13 @@ class TfqInnerProductGradOp : public tensorflow::OpKernel {
           // if applicable compute control qubit mask and control value bits.
           uint64_t mask = 0;
           uint64_t cbits = 0;
-          for (size_t k = 0; k < cur_gate.controlled_by.size(); k++) {
+          for (int k = 0; k < cur_gate.controlled_by.size(); k++) {
             uint64_t control_loc = cur_gate.controlled_by[k];
             mask |= uint64_t{1} << control_loc;
             cbits |= ((cur_gate.cmask >> k) & 1) << control_loc;
           }
 
-          for (size_t k = 0;
+          for (int k = 0;
                k < gradient_gates[cur_batch_index][l - 1].grad_gates.size();
                k++) {
             // Copy sv_adj onto scratch2 in anticipation of non-unitary
